+++ src/bl4ckscor3/bot/bl4ckb0t/misc/LinkTitle.java
@@ -42,6 +42,8 @@ public class LinkTitle
 					GitHub.showCommit(event, s);
 				else if(Core.bot.getConfig().isEnabled("showGitHubRepoInfo") && s.contains("github.com"))
 					GitHub.showRepo(event, s);
+				else if(Core.bot.getConfig().isEnabled("kickOnBannedImgurLink") && event.getChannel().getName().equals("#bl4ckscor3") && (s.contains("imgur") && !s.contains("i.imgur")))
+					Core.bot.kick(event.getChannel().getName(), event.getUser().getNick(), "Only use i.imgur.com links.");
 				else if(Core.bot.getConfig().isEnabled("showLinkTitles"))
 				{
 					WebDriver driver = new HtmlUnitDriver();